requirements.txt
django
djangorestframework
djangorestframework-simplejwt
PyYAML
requests
django-cors-headers